ai_agent/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-core==0.1.35
langchain-community==0.0.21
langchain-openai==0.0.6
openai==1.12.0
jinja2==3.1.3
reportlab==4.1.0
python-multipart==0.0.9
typing-extensions==4.9.0
